Refactor extended MIPS analyzer to deduplicate code.
The initial implementation was just a copy of the original analyzer, but with additional logic for CP propagation. Now,
the original MIPS analyzer is simply extended. A new option was also added to enable/disable CP propagation.

Since this custom analyzer has the same name as the official MIPS analyzer, it seems to have the effect of replacing it.
This seems appropriate, since it only augments the official analyzer; it's a superset in terms of functionality.

One current issue with this implementation is that a single run of the analyzer will not flow all chained functions to their
very end, meaning not all functions references through the CP will be discovered. This can be worked around by simply running
the analyzer repeatedly until the number of discovered functions stops increasing. I've yet to investigate the root cause.